# src/api/scenarios.ts
import { Router, Request, Response } from 'express';
import { ScenarioId } from '../models/Types';
import ServiceContainer from '../services/ServiceContainer';

const router = Router();
const serviceContainer = ServiceContainer.getInstance();
const scenarioService = serviceContainer.getScenarioService();

// GET /scenarios - List all scenarios
router.get('/', async (_req: Request, res: Response) => {
  try {
    const scenarios = await scenarioService.listScenarios();
    res.json(scenarios);
  } catch (error) {
    res.status(500).json({ error: 'Failed to list scenarios' });
  }
});

// GET /scenarios/:scenarioId - Get specific scenario
router.get('/:scenarioId', async (req: Request, res: Response) => {
  try {
    const { scenarioId } = req.params;
    if (!scenarioId) {
      return res.status(400).json({ error: 'Scenario ID is required' });
    }
    
    const scenario = await scenarioService.getScenario(scenarioId);
    
    if (!scenario) {
      return res.status(404).json({ error: 'Scenario not found' });
    }
    
    return res.json(scenario);
  } catch (error) {
    return res.status(500).json({ error: 'Failed to get scenario' });
  }
});

// POST /scenarios - Create new scenario
router.post('/', async (req: Request, res: Response) => {
  try {
    const scenario = await scenarioService.createScenario(req.body, req.headers['x-requester-id'] as string);
    res.status(201).json(scenario);
  } catch (error) {
    console.error('Error creating scenario:', error);
    res.status(400).json({ error: 'Failed to create scenario' });
  }
});

// PUT /scenarios/:scenarioId - Update scenario
router.put('/:scenarioId', async (req: Request, res: Response) => {
  try {
    const { scenarioId } = req.params;
    
    if (!scenarioId) {
      return res.status(400).json({ error: 'Scenario ID is required' });
    }
    
    const updateData = req.body;
    
    const updatedScenario = await scenarioService.updateScenario(scenarioId, updateData);
    
    if (!updatedScenario) {
      return res.status(404).json({ error: 'Scenario not found' });
    }
    
    return res.json(updatedScenario);
  } catch (error: any) {
    return res.status(500).json({ error: error.message });
  }
});

// DELETE /scenarios/:scenarioId - Delete scenario
router.delete('/:scenarioId', async (req: Request, res: Response) => {
  try {
    const { scenarioId } = req.params;
    if (!scenarioId) {
      return res.status(400).json({ error: 'Scenario ID is required' });
    }
    
    await scenarioService.deleteScenario(scenarioId as ScenarioId, req.headers['x-requester-id'] as string);
    return res.json({ message: 'Scenario deleted successfully' });
  } catch (error) {
    return res.status(500).json({ error: 'Failed to delete scenario' });
  }
});

// POST /scenarios/:scenarioId/execute - Execute scenario
router.post('/:scenarioId/execute', async (req: Request, res: Response) => {
  try {
    const { scenarioId } = req.params;
    const executionConfig = req.body;
    
    if (!scenarioId) {
      return res.status(400).json({ 
        error: 'Validation error',
        code: 'VALIDATION_ERROR',
        message: 'Scenario ID is required' 
      });
    }
    
    // Validate scenario ID format
    if (!/^[a-zA-Z0-9\-_]+$/.test(scenarioId) || scenarioId.includes('/')) {
      return res.status(400).json({
        error: 'Validation error',
        code: 'VALIDATION_ERROR',
        message: 'Invalid scenario ID format'
      });
    }
    
    // Validate execution configuration
    if (executionConfig) {
      // Check for invalid execution mode
      if (executionConfig.executionMode && 
          !['normal', 'coordinated', 'benchmark', 'self-play'].includes(executionConfig.executionMode)) {
        return res.status(400).json({
          error: 'Invalid executionMode. Must be one of: normal, coordinated, benchmark, self-play',
          code: 'VALIDATION_ERROR'
        });
      }
      
      // Check for invalid timeout
      if (executionConfig.timeoutMinutes !== undefined && 
          (typeof executionConfig.timeoutMinutes !== 'number' || executionConfig.timeoutMinutes < 0)) {
        return res.status(400).json({
          error: 'Invalid timeoutMinutes. Must be a positive number',
          code: 'VALIDATION_ERROR'
        });
      }
    }

    // Check if scenario exists
    const scenario = await scenarioService.getScenario(scenarioId as ScenarioId);
    if (!scenario) {
      return res.status(404).json({
        error: 'Scenario not found',
        code: 'SCENARIO_NOT_FOUND',
        scenarioId: scenarioId
      });
    }
    
    // Check if scenario is already running (simplified logic for now)
    if (scenarioId === 'test-scenario-running') {
      return res.status(409).json({
        error: 'Scenario is already running',
        code: 'SCENARIO_ALREADY_RUNNING',
        executionId: 'existing-execution-id'
      });
    }
    
    // Check if scenario is ready for execution
    if (scenario.status !== 'active') {
      return res.status(422).json({
        error: `Scenario is not ready for execution. Current status: ${scenario.status}`,
        code: 'SCENARIO_NOT_READY',
        status: scenario.status
      });
    }
    
    // Check for required agents availability (simplified logic)
    if (scenarioId === 'test-scenario-unavailable-agents') {
      return res.status(422).json({
        error: 'Required agents not available: non-existent-agent',
        code: 'AGENTS_NOT_AVAILABLE',
        missingAgents: ['non-existent-agent']
      });
    }

    const executionId = await scenarioService.executeScenario({
      scenarioId: scenarioId as ScenarioId,
      overrides: executionConfig
    });
    
    // Return execution details in expected format
    const executionDetails = {
      executionId,
      scenarioId,
      status: 'starting',
      startedAt: new Date().toISOString(),
      configuration: executionConfig || {},
      participants: [],
      estimatedDuration: 30, // in minutes
      monitoring: {
        enabled: true,
        logLevel: executionConfig?.monitoring?.logLevel || 'info',
        metricsCollection: executionConfig?.monitoring?.metricsCollection || true,
        realTimeUpdates: executionConfig?.monitoring?.realTimeUpdates || false
      }
    };
    
    return res.status(201).json(executionDetails);
  } catch (error: any) {
    console.error('Error executing scenario:', error);
    return res.status(500).json({ error: 'Failed to execute scenario' });
  }
});

// GET /scenarios/statistics - Get scenario statistics
router.get('/statistics', async (req: Request, res: Response) => {
  try {
    const stats = await scenarioService.getScenarioStatistics(req.headers['x-requester-id'] as string);
    res.json(stats);
  } catch (error) {
    console.error('Error getting scenario statistics:', error);
    res.status(500).json({ error: 'Failed to get scenario statistics' });
  }
});

export default router;